requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
auto-py-to-exe==2.6.6
barcode==1.0.1
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.1
distlib==0.3.0
Django==3.1.3
django-utils==0.0.2
Eel==0.11.0
filelock==3.0.12
fpdf==1.7.2
future==0.18.2
gevent==20.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.10
mysql-connector-python==8.0.19
pefile==2019.4.18
Pillow==8.0.0
pipenv==2020.11.15
psycopg2==2.8.5
pycparser==2.20
pygame==1.9.6
PyInstaller==3.6
pynput==1.6.8
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyqt5-tools==5.14.2.1.7b3
python-dotenv==0.13.0
python-Levenshtein==0.12.0
pytz==2020.4
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.10
virtualenv==20.0.20
virtualenv-clone==0.5.4
whichcraft==0.6.1
